# Remove fmt.Sprintf call in attribute Render