@oderwat 2020-10-29
Interesting. Did you try that with something else than `%v` which probably has overhead? No experience here with how the Go compiler has implemented this. I am currently "fighting" something similar in a PHP project, where `"$name = $data"` seems to be better optimized than `$name." = ".$data`. At least for multiple fields, because the `""` version allocates only the final size and copies the parts into it.