@markuswustenberg 2020-10-29
Yeah, I tried %s as well, but that was the same as %v. 

Normally I wouldn't micro-optimise like this, but given that this will be called a lot, it seemed worth it. 

I'm playing with a version of Render that doesn't build everything in-memory, by the way, but uses io.Writer instead. I still want to keep a simple API, so haven't decided yet.